Fix a race condition when fetching args in a Rpc Proc
If [call] is called on the same Proc simulatenously more than once,
there is sometimes a race condition where we try to fetch and set
args when the args have already been set, resulting in the error
"once cell set twice".

This fixes the issue by just ignoring the result of OnceCell::set.