add timeout for wsproxy
Sometimes when network error, the warp lib does not notice, and the reply() function will block, affecting the whole program, making the netidx subscription very slow. So I add a timeout here. If reached the timeout, we just break the loop and close that websocket connection, so it does not affect the rest of the clients.